+++ src/EnvironmentPath.php
@@ -16,11 +16,13 @@ final class EnvironmentPath
     /**
      * @psalm-pure
      */
+    #[\NoDiscard]
     public static function of(string $value): self
     {
         return new self($value);
     }
 
+    #[\NoDiscard]
     public function toString(): string
     {
         return $this->value;