+++ src/Server/Process/User.php
@@ -30,6 +30,7 @@ final class User
     /**
      * @return non-empty-string
      */
+    #[\NoDiscard]
     public function toString(): string
     {
         return $this->value;