+++ src/Server/Processes/Logger.php
@@ -3,22 +3,24 @@ declare(strict_types = 1);
 
 namespace Innmind\Server\Status\Server\Processes;
 
-use Innmind\Server\Status\{
-    Server\Processes,
-    Server\Process,
-    Server\Process\Pid,
+use Innmind\Server\Status\Server\{
+    Process,
+    Process\Pid,
 };
-use Innmind\TimeContinuum\Format;
+use Innmind\Time\Format;
 use Innmind\Immutable\{
-    Set,
+    Sequence,
     Maybe,
 };
 use Psr\Log\LoggerInterface;
 
-final class Logger implements Processes
+/**
+ * @internal
+ */
+final class Logger implements Implementation
 {
     private function __construct(
-        private Processes $processes,
+        private Implementation $processes,
         private LoggerInterface $logger,
     ) {
     }
@@ -26,13 +28,13 @@ final class Logger implements Processes
     /**
      * @internal
      */
-    public static function of(Processes $processes, LoggerInterface $logger): self
+    public static function of(Implementation $processes, LoggerInterface $logger): self
     {
         return new self($processes, $logger);
     }
 
     #[\Override]
-    public function all(): Set
+    public function all(): Sequence
     {
         $all = $this->processes->all();
         $this->logger->debug('{count} processes currently running', [