+++ tests/Facade/Memory/LinuxFacadeTest.php
@@ -8,8 +8,10 @@ use Innmind\Server\Status\{
     Server\Memory,
 };
 use Innmind\Server\Control\ServerFactory as Control;
-use Innmind\TimeContinuum\Clock;
-use Innmind\TimeWarp\Halt\Usleep;
+use Innmind\Time\{
+    Clock,
+    Halt,
+};
 use Innmind\IO\IO;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
@@ -22,7 +24,7 @@ class LinuxFacadeTest extends TestCase
         $this->server = Control::build(
             Clock::live(),
             IO::fromAmbientAuthority(),
-            Usleep::new(),
+            Halt::new(),
         );
     }